modified datasets and model to ensure types of labels and preds are float

+++ pose_est_nets/datasets/datasets.py
@@ -26,7 +26,7 @@ class TrackingDataset(torch.utils.data.Dataset):
         """
         csv_data = pd.read_csv(root_directory + csv_path, header=header_rows)
         self.image_names = list(csv_data.iloc[:,0])
-        self.labels = torch.tensor(csv_data.iloc[:, 1:].to_numpy(), dtype=torch.float64)
+        self.labels = torch.tensor(csv_data.iloc[:, 1:].to_numpy(), dtype=torch.float32)
         self.transform = transform
         self.root_directory = root_directory
         self.num_targets = self.labels.shape[1]
